+++ PeopleWith/Views/Medications/SingleMedication.xaml.cs
@@ -10,7 +10,23 @@ public partial class SingleMedication : ContentPage
     ObservableCollection<MedtimesDosages> Schedule = new ObservableCollection<MedtimesDosages>();
     usermedication MedSelected = new usermedication();
     string[] freqSplit;
+    //Connectivity Changed 
+    public event EventHandler<bool> ConnectivityChanged;
+    //Crash Handler
     CrashDetected crashHandler = new CrashDetected();
+
+    async public void NotasyncMethod(Exception Ex)
+    {
+        try
+        {
+            await crashHandler.CrashDetectedSend(Ex);
+        }
+        catch (Exception ex)
+        {
+            //Dunno 
+        }
+    }
+
 
     //public SingleMedication()
     //{
@@ -120,23 +136,35 @@ public partial class SingleMedication : ContentPage
            
             ScheduleTimes.ItemsSource = Schedule; 
 
-
         }
         catch (Exception Ex)
         {
-
+            NotasyncMethod(Ex);
         }
-
     }
 
     async private void Button_Clicked(object sender, EventArgs e)
     {
         try
         {
-            await Navigation.PushAsync(new MainSchedule(), false);
+            //Connectivity Changed 
+            NetworkAccess accessType = Connectivity.Current.NetworkAccess;
+            if (accessType == NetworkAccess.Internet)
+            {
+                //Limit No. of Taps 
+                ScheduleBtn.IsEnabled = false;
+                await Navigation.PushAsync(new MainSchedule(), false);
+                ScheduleBtn.IsEnabled = true;
+            }
+            else
+            {
+                var isConnected = accessType == NetworkAccess.Internet;
+                ConnectivityChanged?.Invoke(this, isConnected);
+            }
         }
-        catch 
-        { 
+        catch (Exception Ex)
+        {
+            NotasyncMethod(Ex);
         }
     }
 
@@ -144,57 +172,83 @@ public partial class SingleMedication : ContentPage
     {
         try
         {
-            await Navigation.PushAsync(new ShowAllMedication(MedSelected), false); 
+            //Connectivity Changed 
+            NetworkAccess accessType = Connectivity.Current.NetworkAccess;
+            if (accessType == NetworkAccess.Internet)
+            {
+                //Limit No. of Taps 
+                showallbtn.IsEnabled = false;
+                await Navigation.PushAsync(new ShowAllMedication(MedSelected), false);
+                showallbtn.IsEnabled = true;
+            }
+            else
+            {
+                var isConnected = accessType == NetworkAccess.Internet;
+                ConnectivityChanged?.Invoke(this, isConnected);
+            }
         }
-        catch
+        catch (Exception Ex)
         {
+            NotasyncMethod(Ex);
         }
     }
 
-   async private void DeleteBtn_Clicked(object sender, EventArgs e)
+    async private void DeleteBtn_Clicked(object sender, EventArgs e)
     {
         //Delete Medication 
         try
         {
-            bool Result = await DisplayAlert("Delete Medication", "Are you sure you would like to Delete this Medicaiton, it cannot be retrieved once Deleted", "Delete", "Cancel");
-            if (Result)
+            //Connectivity Changed 
+            NetworkAccess accessType = Connectivity.Current.NetworkAccess;
+            if (accessType == NetworkAccess.Internet)
             {
-                //Delete
-                MedSelected.deleted = true; 
-
-                APICalls database = new APICalls();
-                await database.DeleteMedication(MedSelected);
-
-                //Symptom Deleted Message
-                await MopupService.Instance.PushAsync(new PopupPageHelper("Medication Deleted") { });
-                await Task.Delay(1500);
-
-
-                await MopupService.Instance.PopAllAsync(false);
-
-
-                UserMedications.Remove(MedSelected); 
-
-                await Navigation.PushAsync(new AllMedications(UserMedications));
-                var pageToRemoves = Navigation.NavigationStack.FirstOrDefault(p => p is AllMedications);
-                if (pageToRemoves != null)
+                //Limit No. of Taps 
+                DeleteBtn.IsEnabled = false;
+                bool Result = await DisplayAlert("Delete Medication", "Are you sure you would like to Delete this Medicaiton, it cannot be retrieved once Deleted", "Delete", "Cancel");
+                if (Result)
                 {
-                    Navigation.RemovePage(pageToRemoves);
+                    //Delete
+                    MedSelected.deleted = true;
+
+                    APICalls database = new APICalls();
+                    await database.DeleteMedication(MedSelected);
+
+                    //Symptom Deleted Message
+                    await MopupService.Instance.PushAsync(new PopupPageHelper("Medication Deleted") { });
+                    await Task.Delay(1500);
+
+
+                    await MopupService.Instance.PopAllAsync(false);
+
+
+                    UserMedications.Remove(MedSelected);
+
+                    await Navigation.PushAsync(new AllMedications(UserMedications));
+                    DeleteBtn.IsEnabled = true;
+                    var pageToRemoves = Navigation.NavigationStack.FirstOrDefault(p => p is AllMedications);
+                    if (pageToRemoves != null)
+                    {
+                        Navigation.RemovePage(pageToRemoves);
+                    }
+                    Navigation.RemovePage(this);
                 }
-                Navigation.RemovePage(this);
+                else
+                {
+                    //Cancel
+                    DeleteBtn.IsEnabled = true;
+                    return;
+                }
+                
             }
             else
             {
-                //Cancel
-                return; 
+                var isConnected = accessType == NetworkAccess.Internet;
+                ConnectivityChanged?.Invoke(this, isConnected);
             }
-          
-
-
         }
         catch (Exception Ex)
         {
-            //await crashHandler.CrashDetectedSend(Ex);
+            NotasyncMethod(Ex);
         }
     }
 
@@ -202,9 +256,21 @@ public partial class SingleMedication : ContentPage
     {
         try
         {
-            MedSelected.EditMedSection = "Details";
-            await Navigation.PushAsync(new AddMedication(UserMedications, MedSelected), false);
-
+            //Connectivity Changed 
+            NetworkAccess accessType = Connectivity.Current.NetworkAccess;
+            if (accessType == NetworkAccess.Internet)
+            {
+                //Limit No. of Taps 
+                EditMed.IsEnabled = false;
+                MedSelected.EditMedSection = "Details";
+                await Navigation.PushAsync(new AddMedication(UserMedications, MedSelected), false);
+                EditMed.IsEnabled = true;
+            }
+            else
+            {
+                var isConnected = accessType == NetworkAccess.Internet;
+                ConnectivityChanged?.Invoke(this, isConnected);
+            }
             //string action = await DisplayActionSheet("Edit Medication", "Cancel", null, "Details", "Schedule");
 
             //if (action == "Details")
@@ -220,7 +286,7 @@ public partial class SingleMedication : ContentPage
         }
         catch (Exception Ex)
         {
+            NotasyncMethod(Ex);
         }
-
     }
 }
